+++ test/sink/connection/client_connection_test.exs
@@ -28,6 +28,7 @@ defmodule Sink.Connection.ClientConnectionTest do
       start_time: 0
     }
   }
+  @unix_now 1_618_150_125
 
   setup :set_mox_from_context
   setup :verify_on_exit!
@@ -39,7 +40,7 @@ defmodule Sink.Connection.ClientConnectionTest do
       offset = 42
       event_data = <<9, 8, 7>>
       message_id = 1234
-      payload = Protocol.encode_payload(:publish, {event_type_id, key, offset, event_data})
+      payload = Protocol.encode_payload(:publish, {event_type_id, key, offset, @unix_now, event_data})
       encoded_message = Protocol.encode_frame(:publish, message_id, payload)
 
       @mod_transport
@@ -52,6 +53,7 @@ defmodule Sink.Connection.ClientConnectionTest do
       @handler
       |> expect(:handle_publish, 1, fn {^event_type_id, ^key},
                                        ^offset,
+                                       @unix_now,
                                        ^event_data,
                                        ^message_id ->
         :ack
@@ -71,7 +73,7 @@ defmodule Sink.Connection.ClientConnectionTest do
       offset = 42
       event_data = <<9, 8, 7>>
       message_id = 1234
-      payload = Protocol.encode_payload(:publish, {event_type_id, key, offset, event_data})
+      payload = Protocol.encode_payload(:publish, {event_type_id, key, offset, @unix_now, event_data})
       encoded_message = Protocol.encode_frame(:publish, message_id, payload)
 
       # expect a NACK
@@ -90,6 +92,7 @@ defmodule Sink.Connection.ClientConnectionTest do
       @handler
       |> expect(:handle_publish, 1, fn {^event_type_id, ^key},
                                        ^offset,
+                                       @unix_now,
                                        ^event_data,
                                        ^message_id ->
         raise(ArgumentError, message: "boom")