+++ test/sink/event_log/ecto_event_client_log_test.exs
@@ -8,6 +8,7 @@ defmodule Sink.EventLog.EctoClientEventLogTest do
   @event_key <<1, 2, 3>>
   @event_data <<0, 0, 0>>
   @offset 1
+  @timestamp 1_618_150_125
 
   setup do
     # Explicitly get a connection before each test
@@ -20,18 +21,18 @@ defmodule Sink.EventLog.EctoClientEventLogTest do
                TestEctoClientEventLog.log(
                  {@client_id, @event_type_id, @event_key},
                  @offset,
-                 @event_data
+                 {@event_data, @timestamp}
                )
 
-      assert @event_data ==
+      assert {@event_data, @timestamp} ==
                TestEctoClientEventLog.get({@client_id, @event_type_id, @event_key}, @offset)
     end
 
     test "gets latest" do
-      assert :ok = TestEctoClientEventLog.log({@client_id, @event_type_id, @event_key}, 1, <<1>>)
-      assert :ok = TestEctoClientEventLog.log({@client_id, @event_type_id, @event_key}, 2, <<2>>)
+      assert :ok = TestEctoClientEventLog.log({@client_id, @event_type_id, @event_key}, 1, {<<1>>, @timestamp})
+      assert :ok = TestEctoClientEventLog.log({@client_id, @event_type_id, @event_key}, 2, {<<2>>, @timestamp})
 
-      assert {2, <<2>>} ==
+      assert {2, <<2>>, @timestamp} ==
                TestEctoClientEventLog.get_latest({@client_id, @event_type_id, @event_key})
     end
   end
@@ -42,7 +43,7 @@ defmodule Sink.EventLog.EctoClientEventLogTest do
                TestEctoClientEventLog.check_dupe(
                  {@client_id, @event_type_id, @event_key},
                  1,
-                 <<2>>
+                 {<<2>>, @timestamp}
                )
     end
 
@@ -51,27 +52,39 @@ defmodule Sink.EventLog.EctoClientEventLogTest do
                TestEctoClientEventLog.log(
                  {@client_id, @event_type_id, @event_key},
                  1,
-                 @event_data
+                 {@event_data, @timestamp}
                )
 
       assert {:ok, true} =
                TestEctoClientEventLog.check_dupe(
                  {@client_id, @event_type_id, @event_key},
                  1,
-                 @event_data
+                 {@event_data, @timestamp}
                )
     end
 
-    test "is {:error, :data_mismatch, _data} for a record that exists" do
+    test "is {:error, :data_mismatch, _data} for a record that exists with different event_data" do
       assert :ok =
                TestEctoClientEventLog.log(
                  {@client_id, @event_type_id, @event_key},
                  1,
-                 @event_data
+                 {@event_data, @timestamp}
                )
 
-      assert {:error, :data_mismatch, @event_data} =
-               TestEctoClientEventLog.check_dupe({@client_id, @event_type_id, @event_key}, 1, 2)
+      assert {:error, :data_mismatch, {@event_data, @timestamp}} =
+               TestEctoClientEventLog.check_dupe({@client_id, @event_type_id, @event_key}, 1, {2, @timestamp})
+    end
+
+    test "is {:error, :data_mismatch, _data} for a record that exists with different timestamp" do
+      assert :ok =
+               TestEctoClientEventLog.log(
+                 {@client_id, @event_type_id, @event_key},
+                 1,
+                 {@event_data, @timestamp}
+               )
+
+      assert {:error, :data_mismatch, {@event_data, @timestamp}} =
+               TestEctoClientEventLog.check_dupe({@client_id, @event_type_id, @event_key}, 1, {@event_data, 2})
     end
   end
 end