+++ test/sink/event_subscription/ecto_generic_event_subscription_test.exs
@@ -9,6 +9,7 @@ defmodule Sink.EventSubscription.EctoEventGenericSubscriptionTest do
   @topic {@event_type_id, @event_key}
   @event_data <<0, 0, 0>>
   @offset 1
+  @timestamp 1_618_150_125
 
   setup do
     # Explicitly get a connection before each test
@@ -17,20 +18,20 @@ defmodule Sink.EventSubscription.EctoEventGenericSubscriptionTest do
 
   describe "with no client_id" do
     test "subscribes to a topic with an event" do
-      :ok = TestEctoGenericEventLog.log(@topic, @offset, @event_data)
+      :ok = TestEctoGenericEventLog.log(@topic, @offset, {@event_data, @timestamp})
 
       assert :ok = TestEctoGenericEventSubscription.subscribe(@topic, @offset)
     end
 
     test "get_offsets for an existing subscription" do
-      :ok = TestEctoGenericEventLog.log(@topic, @offset, @event_data)
+      :ok = TestEctoGenericEventLog.log(@topic, @offset, {@event_data, @timestamp})
       :ok = TestEctoGenericEventSubscription.subscribe(@topic, @offset)
 
       assert {0, 1} == TestEctoGenericEventSubscription.get_offsets(@topic)
     end
 
     test "ack event" do
-      :ok = TestEctoGenericEventLog.log(@topic, @offset, @event_data)
+      :ok = TestEctoGenericEventLog.log(@topic, @offset, {@event_data, @timestamp})
       :ok = TestEctoGenericEventSubscription.subscribe(@topic, @offset)
 
       assert :ok == TestEctoGenericEventSubscription.ack(@topic, @offset)
@@ -39,7 +40,7 @@ defmodule Sink.EventSubscription.EctoEventGenericSubscriptionTest do
     end
 
     test "update_or_create (update)" do
-      :ok = TestEctoGenericEventLog.log(@topic, @offset, @event_data)
+      :ok = TestEctoGenericEventLog.log(@topic, @offset, {@event_data, @timestamp})
 
       assert :ok == TestEctoGenericEventSubscription.update_or_create(@topic, @offset)
       assert {0, 1} == TestEctoGenericEventSubscription.get_offsets(@topic)
@@ -52,7 +53,7 @@ defmodule Sink.EventSubscription.EctoEventGenericSubscriptionTest do
     end
 
     test "event_queue (with a record)" do
-      :ok = TestEctoGenericEventLog.log(@topic, @offset, @event_data)
+      :ok = TestEctoGenericEventLog.log(@topic, @offset, {@event_data, @timestamp})
       :ok = TestEctoGenericEventSubscription.subscribe(@topic, @offset)
 
       assert [{@event_type_id, @event_key, 0, 1}] == TestEctoGenericEventSubscription.queue()
@@ -62,10 +63,10 @@ defmodule Sink.EventSubscription.EctoEventGenericSubscriptionTest do
       :ok = Sink.TestEctoEventTypeConfig.reset()
       # event_type_id_order = [3,2,1,4]
       # event_type_id_order is defined in TestEctoEventTypeConfig
-      :ok = TestEctoGenericEventLog.log({1, <<>>}, 1, <<>>)
-      :ok = TestEctoGenericEventLog.log({2, <<>>}, 1, <<>>)
-      :ok = TestEctoGenericEventLog.log({3, <<>>}, 1, <<>>)
-      :ok = TestEctoGenericEventLog.log({4, <<>>}, 1, <<>>)
+      :ok = TestEctoGenericEventLog.log({1, <<>>}, 1, {<<>>, @timestamp})
+      :ok = TestEctoGenericEventLog.log({2, <<>>}, 1, {<<>>, @timestamp})
+      :ok = TestEctoGenericEventLog.log({3, <<>>}, 1, {<<>>, @timestamp})
+      :ok = TestEctoGenericEventLog.log({4, <<>>}, 1, {<<>>, @timestamp})
       :ok = TestEctoGenericEventSubscription.subscribe({1, <<>>}, 1)
       :ok = TestEctoGenericEventSubscription.subscribe({2, <<>>}, 1)
       :ok = TestEctoGenericEventSubscription.subscribe({3, <<>>}, 1)